vector_db_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vector-db-pipeline"
version = "0.1.0"
description = "Multimodal ingestion and retrieval pipeline for Korean exam question images."
readme = "../README.md"
requires-python = ">=3.10"
authors = [
    { name = "Sohn" },
]
dependencies = [
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "tqdm>=4.65",
    "requests>=2.31",
    "openai>=1.0",
    "transformers>=4.48",
    "torch>=2.4",
    "qdrant-client>=1.9",
    "accelerate>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.9",
    "pre-commit>=3.0",
    "bitsandbytes>=0.43",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort
    "W",     # pycodestyle warnings
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "E501",   # line length (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
